spark/requirements.txt
isodate==0.6.0
joblib==0.17.0
numpy==1.19.4
pandas==1.1.4
parsedatetime==2.6
progressbar2==3.53.1
python-dateutil==2.8.1
pytimeparse==1.1.8
pytz==2020.4
PyYAML==5.3.1
text-unidecode==1.3
tqdm==4.53.0
tx-functional==0.1.2
